+++ cli/cli/commands/service/service.go
@@ -7,7 +7,9 @@ package service
 
 import (
 	"github.com/kurtosis-tech/kurtosis-cli/cli/command_str_consts"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/service/add"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/service/logs"
+	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/service/rm"
 	"github.com/kurtosis-tech/kurtosis-cli/cli/commands/service/shell"
 	"github.com/spf13/cobra"
 )
@@ -19,6 +21,8 @@ var ServiceCmd = &cobra.Command{
 }
 
 func init() {
+	ServiceCmd.AddCommand(add.ServiceAddCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(logs.LogsCmd)
+	ServiceCmd.AddCommand(rm.ServiceRmCmd.MustGetCobraCommand())
 	ServiceCmd.AddCommand(shell.ShellCmd)
 }