+++ cli/docs/changelog.md
@@ -6,6 +6,18 @@
 ### Changes
 * Upgraded the kurtosis core to 1.40.1
 * Upgraded the kurtosis engine to 1.12.0
+* Switched `enclave dump` to use `KurtosisBackend`
+
+# 0.11.9
+### Features
+* Added the `service add` command for adding a service to an enclave
+* Added the `service rm` command for removing services from an enclave
+* Added an `--id` flag to `enclave new` to allow setting the new enclave's ID
+* Users can now provide an existing enclave to the `--enclave-id` parameter of `module exec` to exec the module in an existing enclave, rather than creating a new one
+
+# 0.11.8
+### Fixes
+* Fixed break in the `sandbox` command when we switched to the newer `EnclaveContext.newGrpcNodeEnclaveContext` function
 
 # 0.11.7
 ### Features